fix : aliminf compleated but Edit Code To develop pull
