+++ src/main/java/com/example/finalproj/ml/AlimML/AlimMLService.java
@@ -1,6 +1,5 @@
-package com.example.finalproj.ml.service;
+package com.example.finalproj.ml.AlimML;
 
-import com.example.finalproj.Book.entity.Book;
 import com.example.finalproj.Alim.entity.Alim;
 import com.example.finalproj.AlimInf.entity.AlimInf;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -16,49 +15,15 @@ import java.time.LocalDateTime;
 import java.util.Map;
 
 @Service
-public class MLService {
+public class AlimMLService {
     @Value("${ml.service.url}")
     private String mlServiceUrl;
     private final RestTemplate restTemplate;
     private final ApplicationEventPublisher eventPublisher;
-    private final ObjectMapper objectMapper;
 
-    public MLService(RestTemplate restTemplate, ApplicationEventPublisher eventPublisher, ObjectMapper objectMapper) {
+    public AlimMLService(RestTemplate restTemplate, ApplicationEventPublisher eventPublisher) {
         this.restTemplate = restTemplate;
         this.eventPublisher = eventPublisher;
-        this.objectMapper = objectMapper;
-    }
-
-    // 이미지를 ML 서비스로 전송
-    public void sendImageToMLService(String imagePath, Integer userId, Integer babyId, Integer calendarPhotoId) {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_JSON);
-        String requestBody = String.format("{\"image_path\": \"%s\", \"user_id\": %d, \"baby_id\": %d}",
-                imagePath, userId, babyId);
-        HttpEntity<String> request = new HttpEntity<>(requestBody, headers);
-        try {
-            String mlResponse = restTemplate.postForObject(mlServiceUrl + "/process_image", request, String.class);
-            System.out.println("ML Service response: " + mlResponse);
-            eventPublisher.publishEvent(new CalendarMLProcessingCompletedEvent(this, mlResponse, calendarPhotoId));
-        } catch (Exception e) {
-            System.err.println("Error in ML service process: " + e.getMessage());
-        }
-    }
-
-    // 책 정보를 ML 서비스로 전송
-    public void sendBookToMLService(Book book) {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_JSON);
-        String requestBody = String.format("{\"book_id\": %d, \"user_id\": %d, \"title\": \"%s\", \"cover_path\": \"%s\"}",
-                book.getBookId(), book.getUserId(), book.getTitle(), book.getCoverPath());
-        HttpEntity<String> request = new HttpEntity<>(requestBody, headers);
-        try {
-            String mlResponse = restTemplate.postForObject(mlServiceUrl + "/process_book", request, String.class);
-            System.out.println("ML Service response for book: " + mlResponse);
-            eventPublisher.publishEvent(new BookMLProcessingCompletedEvent(this, mlResponse, book.getBookId()));
-        } catch (Exception e) {
-            System.err.println("Error in ML service process for book: " + e.getMessage());
-        }
     }
 
     // Alim을 ML 서비스로 전송
