+++ src/main/java/com/example/finalproj/ml/CalendarML/CalendarMLEventListener.java
@@ -1,7 +1,5 @@
-package com.example.finalproj.ml.service;
+package com.example.finalproj.ml.CalendarML;
 
-import com.example.finalproj.AlimInf.entity.AlimInf;
-import com.example.finalproj.AlimInf.service.AlimInfService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.event.EventListener;
 import org.springframework.http.HttpEntity;
@@ -12,22 +10,18 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 
 @Component
-public class MLEventListener {
+public class CalendarMLEventListener {
 
     private final RestTemplate restTemplate;
     private final String appUrl;
-    private final AlimInfService alimInfService;
 
     // 생성자
-    public MLEventListener(RestTemplate restTemplate,
-                           @Value("${app.url:http://localhost:8080}") String appUrl,
-                           AlimInfService alimInfService) {
+    public CalendarMLEventListener(RestTemplate restTemplate, @Value("${ml.app.url}") String appUrl) {
         this.restTemplate = restTemplate;
         this.appUrl = appUrl;
-        this.alimInfService = alimInfService;
     }
 
-    // ML 처리 완료 이벤트 핸들러 (기존 코드)
+    // ML 처리 완료 이벤트 핸들러
     @EventListener
     public void handleMLProcessingCompletedEvent(CalendarMLProcessingCompletedEvent event) {
         String mlResponse = event.getMlResponse();
@@ -43,21 +37,6 @@ public class MLEventListener {
             System.out.println("Calendar Photo Inf Controller response: " + forwardResponse.getBody());
         } catch (Exception e) {
             System.err.println("Error forwarding ML response: " + e.getMessage());
-        }
-    }
-
-    // Alim ML 처리 완료 이벤트 핸들러 (새로운 코드)
-    @EventListener
-    public void handleAlimMLProcessingCompletedEvent(AlimMLProcessingCompletedEvent event) {
-        String mlResponse = event.getMlResponse();
-        Integer alimId = event.getAlimId();
-
-        try {
-            AlimInf alimInf = MLService.createAlimInfFromMLResponse(mlResponse, alimId);
-            alimInfService.createAlimInf(alimInf);
-            System.out.println("AlimInf created and saved successfully from event");
-        } catch (Exception e) {
-            System.err.println("Error processing Alim ML response from event: " + e.getMessage());
         }
     }
 }
