+++ src/main/java/com/example/finalproj/ml/CalendarML/CalendarMLProcessingCompletedEvent.java
@@ -1,4 +1,4 @@
-package com.example.finalproj.ml.service;
+package com.example.finalproj.ml.CalendarML;
 
 import org.springframework.context.ApplicationEvent;
